requirements.txt
numpy==2.2.6
scipy==1.15.2
pandas
matplotlib==3.10.5
seaborn
tqdm
scikit-learn
statsmodels
sympy
cvxopt==1.3.2
cvxpy
torch
torch-geometric==2.6.1
rdkit
GraKeL==0.1.10
shapely
python-ternary
reportlab